chore: clean file names, refactor api code
