+++ src/components/Dashboard/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
 import { useGetBookingsQuery } from '../../features/api/apiSlice';
-import { IBooking } from '../../types/types';
+import { IBooking } from '../../types';
 import { Booking, Resource } from '../';
 
 import styles from './style.module.css';
